v0.5 LuaCallable calls methods from LuaController. Better comments.
LuaCallable now, instead of emitting a signal from a LuaController Object, calls a method from the Object, then sends the result as a return value to the call inside Lua.

Changes comments in the files to be more compliant with Godot's documentation style guide.

# LuaController.* :
In `prepare_callables` now uses the list of methods to make the vector, instead of the list of signals.
 - Changes the wording of the module from "signal" to "method"

# LuaCallable.* :
Changed from emiting signals to calling methods:
 - When called inside Lua, now has a return value.
 - Uses `Object::call()` instead of `Object::emit_signal()`.
Because `call` provides a way to check the error that occured with the call through the `Variant::CallError` type, sends that information to the error handler.
 - The integer passed to `lua_error_handler` now represents the error according to the enum `Variant::CallError::Error`, not the enum `Error`