v0.6.1 Configures Doxyfile, Comments corrections, useful README, automated testing
# Doxyfile :
configure it to build the documentation as HTML.

# README.md :
Explains all compilation and installation necessary for the use of the module

# lua_callable.h :
Correction of code comments

# LuaControllerContext.* :
Correction of code comments.
Changes return type for passing tests

# LuaControllerUnitTester.* :
Change return type for unit test to always be Array. If all tests passed, it will be empty.
Fixes TEST 7 assertion for LuaController